Add [].Del and [].Indices
