+++ src/interp/tests/samples/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(basic)
 add_subdirectory(complex)
+add_subdirectory(extra)
