+++ src/interp/tests/samples/basic/CMakeLists.txt
@@ -10,4 +10,4 @@ foreach (file IN LISTS files)
         DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/${file} VERBATIM)
 endforeach()
 add_custom_target(interp_files_basic DEPENDS ${files})
-add_dependencies(interp interp_files_basic)
+add_dependencies(interp_tests interp_files_basic)
