+++ src/interp/tests/samples/complex/CMakeLists.txt
@@ -10,4 +10,4 @@ foreach (file IN LISTS files)
         DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/${file} VERBATIM)
 endforeach()
 add_custom_target(interp_files_complex DEPENDS ${files})
-add_dependencies(interp interp_files_complex)
+add_dependencies(interp_tests interp_files_complex)
